Change getLength to be a default interface method, so removed it from many command classes. Also gave chord command a length. Story: 280

+++ src/main/java/seng302/command/PlayNote.java
@@ -32,7 +32,7 @@ public class PlayNote implements Command {
         this.duration = duration;
     }
 
-    public float getLength(Environment env) {
+    public long getLength(Environment env) {
         return 60000 / env.getPlayer().getTempo();
     }
